% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21088R.R
\name{Addo}
\alias{Addo}
\title{Add column of 1 to data}
\usage{
Addo(data)
}
\arguments{
\item{data}{Data matrix without adding constant 1 column.}
}
\value{
Data with first column filled with 1.  \code{new_data}
}
\description{
Before using the algorithm, the first column of data should be filled with 1.
}
\examples{
\dontrun{
A = matrix(
c(1,1,1,1,1,2,3,4),4,2)
x <- Addo(A)
print(x)
}
}
