% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21088R.R
\name{Cost}
\alias{Cost}
\title{Cost}
\usage{
Cost(input_data, real_result, theta)
}
\arguments{
\item{input_data}{Input_data  matrix after adding constant 1 column}

\item{real_result}{Real_result vector whose length is equal to the column number of data.}

\item{theta}{The parameters of linear regression}
}
\value{
cost \code{cost}
}
\description{
loss function
}
\examples{
\dontrun{
x <- seq(0.1,10,0.002)
n <- length(x)
y <- 2*x+5+rnorm(n)
z <- as.matrix(data.frame(rep(1,n),x))
cost <- Cost(z, y, c(1,1))
print(cost)
}
}
