% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{Rcpp function}
\usage{
gibbsC(N, a, b, n)
}
\arguments{
\item{N}{sample size}

\item{a}{The first parameter of the distribution}

\item{b}{The second parameter of the distribution}

\item{n}{The third parameter of the distribution}
}
\value{
generated random numbers matrix \code{mat}
}
\description{
an Rcpp function of gibbs sample
}
\examples{
\dontrun{
C <- gibbsC(5000,1,1,25)
print(C)
}
}
