% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21088R.R
\name{Shuffle_sequence}
\alias{Shuffle_sequence}
\title{Shuffle sequence of data}
\usage{
Shuffle_sequence(data)
}
\arguments{
\item{data}{Data matrix after adding constant 1 column.}
}
\value{
New sequence  \code{random_sequence}
}
\description{
Before using the algorithm, the data needs to be randomly shuffled to eliminate the impact of the data order.
}
\examples{
\dontrun{
A = matrix(
c(1,1,1,1,1,2,3,4),4,2)
x <- Shuffle_sequence(A)
print(x)
}
}
