% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21088R.R
\name{train_BGD}
\alias{train_BGD}
\title{Using Batch Gradient Descent to train the model.}
\usage{
train_BGD(input_data, real_result, iter, alpha, theta)
}
\arguments{
\item{input_data}{Input_data  matrix after adding constant 1 column}

\item{real_result}{Real_result vector whose length is equal to the column number of data.}

\item{iter}{iterations}

\item{alpha}{Learning rate}

\item{theta}{The initial parameters of linear regression}
}
\value{
theta after iterations and cost after every training \code{(theta ,cost)}
}
\description{
Use Batch Gradient Descent to train the model.
}
\examples{
\dontrun{
x <- seq(0.1,10,0.01)
n <- length(x)
z <- rnorm(n)
y <- 2*x+5+z
bgd <- train_BGD(x,y,200,0.01,c(1,1))
print(bgd)
}
}
