% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21088R.R
\name{train_MBGD}
\alias{train_MBGD}
\title{Using Mini-Batch Gradient Descent to train the model.}
\usage{
train_MBGD(input_data, real_result, iter, batch_size, alpha, theta)
}
\arguments{
\item{input_data}{Input_data  matrix after adding constant 1 column}

\item{real_result}{Real_result vector whose length is equal to the column number of data.}

\item{iter}{iterations}

\item{batch_size}{Batch_size parameter constant}

\item{alpha}{Learning rate}

\item{theta}{The initial parameters of linear regression}
}
\value{
theta after iterations and cost after every iteration \code{(theta ,cost)}
}
\description{
Use Mini-Batch Gradient Descent to train the model.
}
\examples{
\dontrun{
x <- seq(0.1,10,0.01)
n <- length(x)
z <- rnorm(n)
y <- 2*x+5+z
theta <- train_MBGD(x,y,500,200,0.01,c(1,1))$theta
print(theta)
cost <- train_MBGD(x,y,500,200,0.01,c(1,1))$cost
print(cost)
}
}
